Initial Import.
The source check-in is based on U-Boot 1.0.0 (http://u-boot.sourceforge.net/).

It contains board support platform files needed for the MagicEyes MMSP2 ARM based reference board.

This should enable a boot-loader to be constructed for any MMSP2/MP2520F based system.

It is NOT a complete distribution of U-Boot but complete enough to build for the MMSP2/MP2520F.


git-svn-id: svn://svn.code.sf.net/p/open2x/code/trunk@9 3edcfc3d-570e-0410-a036-8bc182068dc2